Update class name TimeLine->BookDetailView(BookDetailView.jsx)
